Move Machine (etc.) and Links into own modules.
Deprecates the `rig.machine` module. Importing this module produces a
deprecation warning but otherwise contains aliases to the original structures.

Moves Machine, Cores, SDRAM and SRAM into rig.place_and_route.machine (aliased
in rig.place_and_route) to better reflect the P&R-specific nature of these
structures.

Moves Links into `rig.links` as a more sensible home.
