+++ docs/source/links.rst
@@ -1,0 +1,7 @@
+:py:mod:`rig.links`: Chip-to-chip link data structure
+-----------------------------------------------------
+
+.. automodule:: rig.links
+    :members:
+    :special-members:
+
