+++ docs/source/machine.rst
@@ -1,6 +1,12 @@
-:py:mod:`rig.machine`: Machine resource data structures
--------------------------------------------------------
+:py:mod:`rig.place_and_route.Machine` etc.: Machine resource availability P&R data structure
+--------------------------------------------------------------------------------------------
 
-.. automodule:: rig.machine
+.. autoclass:: rig.place_and_route.Machine
     :members:
     :special-members:
+
+.. autodata:: rig.place_and_route.Cores
+
+.. autodata:: rig.place_and_route.SDRAM
+
+.. autodata:: rig.place_and_route.SRAM
