+++ tests/place_and_route/allocate/test_generic_allocate.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from rig.machine import Machine, Cores, SDRAM
+from rig.place_and_route.machine import Machine, Cores, SDRAM
 
 from rig.routing_table import Routes
